debugging gouraud shader (it broke!!!!!!!!!!!!)
