+++ src/shaders.rs
@@ -1,11 +1,13 @@
-pub mod shader;
-pub mod gouraud;
-pub mod normal;
-pub mod tangent_normal;
+mod shader;
+mod gouraud;
+mod normal;
+mod normal_spec;
+mod tangent_normal;
 
 pub use shader::Shader;
 pub use gouraud::GouraudShader;
-pub use normal::{NormalMappedShader, NormalSpecularShader};
+pub use normal::NormalMappedShader;
+pub use normal_spec::NormalSpecularShader;
 pub use tangent_normal::TangentNormalShader;
